LEVELS of information
Before going through the whole EFSA model, needed just to set up all
needed levels:

Observation, SAMPLE, surveillance system and activity (an activity NOT
is-a system)

[System and activity - not a IS-A, but a part of. I thougth I could
solve this with naming (hazard specific activity as a sub-group from
hazard specific surveillance). But in the end, these things are going to
be INSTANCES. So we wouldn't have the instance activity being inside
another instance. Better to specifically describe it with a property
(like herd and animal)]

AH: surveillance system and specific activities
PH: probably only system, and case-based

-->
Actually concluded that a surveillance activity is a type of observation
context.
So move surveillance activity and system there, and keep the properties
that connect them, except that generalized for any observation context

matrix of relevant properties to connect all levels documented.

--> ALSO:
Observed in anatomical location --> has anatomical location so that it
can apply also to sample